+++ packages/in_app_purchases_intl/lib/extensions/duration_extension.dart
@@ -1,52 +1,59 @@
 import 'package:flutter/material.dart';
+import 'package:in_app_purchases_intl/delegate/paywall_localizations.dart';
 import 'package:in_app_purchases_intl/helper/paywall_helper.dart';
 import 'package:in_app_purchases_intl/model/duration_strings.dart';
 
 /// Extension to get all strings for a specific duration
 extension DurationExtension on String {
-  /// Saves the translations of a subscription length to a model
-  DurationStrings toDurationStrings(BuildContext context) {
-    final paywallIntl = PaywallL10NHelper.of(context);
-    // Switch the cases we know
-    switch (this.toUpperCase()) {
-      case "P1W":
-        return DurationStrings(
-            durationText: paywallIntl.weekly,
-            durationTextNumber: paywallIntl.nweek(1),
-            durationType: paywallIntl.week,
-            durationNumber: 1);
-      case "P4W":
-        return DurationStrings(
-            durationText: paywallIntl.weekly,
-            durationTextNumber: paywallIntl.nweek(4),
-            durationType: paywallIntl.week,
-            durationNumber: 4);
-      case "P1M":
-        return DurationStrings(
-            durationText: paywallIntl.monthly,
-            durationTextNumber: paywallIntl.nmonth(1),
-            durationType: paywallIntl.month,
-            durationNumber: 1);
-      case "P3M":
-        return DurationStrings(
-            durationText: paywallIntl.quarterly,
-            durationTextNumber: paywallIntl.nmonth(3),
-            durationType: paywallIntl.month,
-            durationNumber: 3);
-      case "P6M":
-        return DurationStrings(
-            durationText: paywallIntl.biannual,
-            durationTextNumber: paywallIntl.nmonth(6),
-            durationType: paywallIntl.month,
-            durationNumber: 6);
-      case "P1Y":
-        return DurationStrings(
-            durationText: paywallIntl.yearly,
-            durationTextNumber: paywallIntl.nyear(1),
-            durationType: paywallIntl.year,
-            durationNumber: 1);
+  @Deprecated("please use [fromIso8601]")
+  DurationStrings toDurationStrings(BuildContext context) => fromIso8601(PaywallL10NHelper.of(context));
+
+  /// ISO Codes such as P1W, P3W, P2Y etc.
+  ///
+  /// You can get the PaywallLocalizations with `PaywallL10NHelper.of(context)`
+  DurationStrings fromIso8601(PaywallLocalizations paywallIntl) {
+    final regex = RegExp(r'(\d+)([A-Z])');
+    final match = regex.firstMatch(this);
+    final amount = int.parse(match!.group(1)!);
+    final unitString = match.group(2);
+
+    if (unitString == null) {
+      throw ArgumentError("String does not have a format of P1D, P1W, P1M, P1Y ");
     }
-    // fallback to a zero object
-    return DurationStrings.zero(paywallIntl);
+    return switch (unitString) {
+      "D" => DurationStrings(
+          durationText: paywallIntl.day,
+          durationTextNumber: paywallIntl.nday(amount),
+          durationType: paywallIntl.day,
+          durationNumber: amount),
+      "W" => DurationStrings(
+          durationText: paywallIntl.weekly,
+          durationTextNumber: paywallIntl.nweek(amount),
+          durationType: paywallIntl.week,
+          durationNumber: amount),
+      "M" => switch (amount) {
+          3 => DurationStrings(
+              durationText: paywallIntl.quarterly,
+              durationTextNumber: paywallIntl.nmonth(amount),
+              durationType: paywallIntl.month,
+              durationNumber: amount),
+          6 => DurationStrings(
+              durationText: paywallIntl.biannual,
+              durationTextNumber: paywallIntl.nmonth(6),
+              durationType: paywallIntl.month,
+              durationNumber: 6),
+          _ => DurationStrings(
+              durationText: paywallIntl.monthly,
+              durationTextNumber: paywallIntl.nmonth(amount),
+              durationType: paywallIntl.month,
+              durationNumber: amount)
+        },
+      "Y" => DurationStrings(
+          durationText: paywallIntl.yearly,
+          durationTextNumber: paywallIntl.nyear(amount),
+          durationType: paywallIntl.year,
+          durationNumber: amount),
+      _ => throw ArgumentError()
+    };
   }
 }