+++ packages/in_app_purchases_intl/lib/model/duration_strings.dart
@@ -1,5 +1,3 @@
-import 'package:in_app_purchases_intl/delegate/paywall_localizations.dart';
-
 class DurationStrings {
   /// e.g. biannual
   final String durationText;
@@ -7,7 +5,7 @@ class DurationStrings {
   /// e.g. 6 months
   final String durationTextNumber;
 
-  /// e.g. month
+  /// e.g. monthly, quarterly, biannual
   final String durationType;
 
   /// e.g. 6
@@ -18,11 +16,4 @@ class DurationStrings {
       required this.durationTextNumber,
       required this.durationType,
       required this.durationNumber});
-
-  /// Empty fallback object
-  DurationStrings.zero(PaywallLocalizations paywallIntl)
-      : durationText = paywallIntl.day,
-        durationNumber = 0,
-        durationTextNumber = paywallIntl.day,
-        durationType = paywallIntl.day;
 }